pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finopspp"
version = "0.0.0"
description = "CLI tool used for the FinOps++ Assessment."
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Victoria Levy" }]
dependencies = [
    "click == 8.2.1",
    "pandas == 2.3.2",
    "pyyaml == 6.0.2",
    "xlsxwriter == 3.2.5",
]

[project.scripts]
finopspp = "finopspp.__main__:cli"

[tool.setuptools]
packages = ["finopspp", "finopspp.controls"]

[tool.setuptools.package-dir]
"finopspp.controls" = "controls"

[tool.setuptools.package-data]
"finopspp.controls" = ["*.yaml"]